requirements.txt
Django
djangorestframework
psycopg2
python-dotenv
djangorestframework-simplejwt
ipython
django-filter
drf-yasg
